* upstream/master: (66 commits)
  Revert egregious lambda-for-the-sake-of-lambda
  String and balance tweaks.
  Better laser-lock warning strings.
  Steel bolt recipe
  Make butchering re-roll skill roll per item
  Russian MOTD fastfix
  Spanish credits update
  Russian update
  Italian update
  Argentinian update
  German update
  Update version name for 0.C release.
  Another typo fix.
  Use submap->field_count to skip empty sm-s
  Optimize rain decaying fire, generalize to other fields
  Post-freeze fixes for effects.json.
  Damn.
  Names tweaks.
  Spacing fixes for post-freeze.
  The boat-board doesn move the boat.
  ...

Conflicts:
	src/itype.h

+++ src/crafting_inventory_t.cpp
@@ -1280,7 +1280,7 @@ int crafting_inventory_t::deserialize(JsonArray &arr, candvec &vec) {
     }
     const int index = arr.get_int(0);
     vec.clear();
-    for(int i = 1; i < arr.size(); i++) {
+    for(size_t i = 1; i < arr.size(); i++) {
         JsonObject o(arr.get_object(i));
         candidate_t ci(*this, o);
         if(!ci.valid()) {
